Move most of tools.cmd package object to tools.partest
